requirements.txt
python-dotenv>=1.0.0
requests>=2.32.0
mcp>=1.9.0 